fix: add explicit float casts to resolve conversion warnings in board_view.cpp

+++ game/src/gui/board_view.cpp
@@ -34,14 +34,14 @@ void BoardView::draw(sf::RenderWindow& window, const Board& board) const {
     
     if (boardTextureLoaded) {
         sf::Sprite boardSprite(boardTexture);
-        boardSprite.setScale(sf::Vector2f(tileSize_ * 8 / boardTexture.getSize().x, tileSize_ * 8 / boardTexture.getSize().y));
+        boardSprite.setScale(sf::Vector2f(tileSize_ * 8 / static_cast<float>(boardTexture.getSize().x), tileSize_ * 8 / static_cast<float>(boardTexture.getSize().y)));
         window.draw(boardSprite);
     } else {
         // Fallback: Draw chessboard squares
         for (int row = 0; row < 8; ++row) {
             for (int col = 0; col < 8; ++col) {
                 sf::RectangleShape square(sf::Vector2f(tileSize_, tileSize_));
-                square.setPosition({col * tileSize_, row * tileSize_});
+                square.setPosition({static_cast<float>(col * tileSize_), static_cast<float>(row * tileSize_)});
                 bool isLight = (row + col) % 2 == 0;
                 square.setFillColor(isLight ? sf::Color(240, 217, 181) : sf::Color(181, 136, 99));
                 window.draw(square);
@@ -64,9 +64,9 @@ void BoardView::draw(sf::RenderWindow& window, const Board& board) const {
                     sf::Sprite pieceSprite(TextureManager::getTexture(imagePath));
                     // Flip the board: row 0 becomes row 7, row 1 becomes row 6, etc.
                     int displayRow = 7 - row;
-                    pieceSprite.setPosition(sf::Vector2f(col * tileSize_, displayRow * tileSize_));
-                    pieceSprite.setScale(sf::Vector2f(tileSize_ / pieceSprite.getTexture().getSize().x, 
-                                                    tileSize_ / pieceSprite.getTexture().getSize().y));
+                    pieceSprite.setPosition(sf::Vector2f(static_cast<float>(col * tileSize_), static_cast<float>(displayRow * tileSize_)));
+                    pieceSprite.setScale(sf::Vector2f(tileSize_ / static_cast<float>(pieceSprite.getTexture().getSize().x), 
+                                                    tileSize_ / static_cast<float>(pieceSprite.getTexture().getSize().y)));
                     window.draw(pieceSprite);
                 } catch (const std::runtime_error& e) {
                     std::cout << "Failed to load piece texture: " << imagePath << " - " << e.what() << std::endl;
@@ -78,8 +78,8 @@ void BoardView::draw(sf::RenderWindow& window, const Board& board) const {
                     // Flip the board for text too
                     int displayRow = 7 - row;
                     pieceText.setPosition(sf::Vector2f(
-                        col * tileSize_ + tileSize_ * 0.2f,
-                        displayRow * tileSize_ + tileSize_ * 0.05f
+                        static_cast<float>(col * tileSize_) + tileSize_ * 0.2f,
+                        static_cast<float>(displayRow * tileSize_) + tileSize_ * 0.05f
                     ));
                     window.draw(pieceText);
                 }
@@ -103,7 +103,7 @@ std::pair<int, int> BoardView::screenToBoard(const sf::Vector2i& mousePos) const
 sf::Vector2f BoardView::boardToScreen(int row, int col) const {
     // Flip the row for display
     int displayRow = 7 - row;
-    return sf::Vector2f(col * tileSize_, displayRow * tileSize_);
+    return sf::Vector2f(static_cast<float>(col * tileSize_), static_cast<float>(displayRow * tileSize_));
 }
 
 std::string BoardView::getImagePath(const Piece& piece) const {
